feat: BME680 routes, socketio
