--- backend/helpers/Readme.md
@@ -1,1 +1,0 @@
-Unfortunately due to licensing, i can't provide the BSEC generic release to github.
