--- backend/helpers/bjon's-lib/bsec_bme680_python-master/make.config
@@ -1,22 +1,0 @@
-BSEC_DIR='./src/BSEC_1.4.6.0_Generic_Release_20180425'
-
-VERSION='Normal_version'
-
-ARCH="${VERSION}/RaspberryPI/PiThree_ArmV8-a-64bits"
-
-# Other architectures can be found in BSEC_DIR/algo/bin/.
-
-CONFIG='generic_33v_3s_4d'
-
-# Other configs are:
-# generic_18v_300s_28d
-# generic_18v_300s_4d
-# generic_18v_3s_28d
-# generic_18v_3s_4d
-# generic_33v_300s_28d
-# generic_33v_300s_4d
-# generic_33v_3s_28d
-# generic_33v_3s_4d
-
-CONFIG_DIR='.'
-
