--- backend/helpers/bme68x-python-library/BME68x-Sensor-API/README.md
@@ -1,30 +1,0 @@
-# BME68X Sensor API
-
-> Bosch Sensortec's BME680 and BME688 sensor API
-
-## Sensor Overview
-BME680 is an integrated environmental sensor developed specifically for mobile applications and wearables where size and low power consumption are key requirements. Expanding Bosch Sensortec’s existing family of environmental sensors, the BME680 integrates for the first time high-linearity and high-accuracy gas, pressure, humidity and temperature sensors. It consists of an 8-pin metal-lid 3.0 x 3.0 x 0.93 mm³ LGA package which is designed for optimized consumption depending on the specific operating mode, long term stability and high EMC robustness. The gas sensor within the BME680 can detect a broad range of gases to measure air quality for personal well being. Gases that can be detected by the BME680 include Volatile Organic Compounds (VOC) from paints (such as formaldehyde), lacquers, paint strippers, cleaning supplies, furnishings, office equipment, glues, adhesives and alcohol.
-
-### Features
-
-- Air quality measurement
-- Personalized weather station
-- Context awareness, e.g. skin moisture detection, room change detection
-- Fitness monitoring / well-being
-- Warning regarding dryness or high temperatures
-- Measurement of volume and air flow
-- Home automation control (e.g. HVAC)
-- GPS enhancement (e.g. time-to-first-fix improvement, dead reckoning, slope detection)
-- Indoor navigation (change of floor detection, elevator detection)
-- Altitude tracking and calories expenditure for sports activities
-
-#### Important links:
-For more information, please refer to: 
-
-- [BME680 Product page](https://www.bosch-sensortec.com/bst/products/all_products/bme680)
-- [BME680 & BME688 Github page](https://github.com/BoschSensortec/BME68x-Sensor-API)
-- [BME680 gas sensor design guide](https://community.bosch-sensortec.com/t5/Knowledge-base/BME680-gas-sensor-series-design-guide/ta-p/5952)
-- [Knowledge base page](https://community.bosch-sensortec.com/t5/Knowledge-base/tkb-p/bst_community-mems-tkb)
-- [Community support page](https://community.bosch-sensortec.com)
-
----
