--- backend/helpers/bme68x-python-library/BME68x-Sensor-API/examples/parallel_mode/Makefile
@@ -1,15 +1,0 @@
-COINES_INSTALL_PATH ?= ../../../..
-
-EXAMPLE_FILE ?= parallel_mode.c
-
-API_LOCATION ?= ../..
-
-C_SRCS += \
-$(API_LOCATION)/bme68x.c \
-../common/common.c
-
-INCLUDEPATHS += \
-$(API_LOCATION) \
-../common
-
-include $(COINES_INSTALL_PATH)/coines.mk
