--- backend/helpers/bme68x-python-library/bsecConstants.py
@@ -1,21 +1,0 @@
-# Enable use of BSEC
-BSEC_ENABLE = 1
-# Disable use of BSEC
-BSEC_DISABLE = 0
-
-# Sample rates
-# Sample rate of a disabled sensor
-BSEC_SAMPLE_RATE_DISABLED = 65535.0
-# Sample rate in case of Ultra Low Power Mode
-BSEC_SAMPLE_RATE_ULP = 0.0033333
-# Sample rate in case of Low Power Mode
-BSEC_SAMPLE_RATE_LP = 0.33333
-# Input value used to trigger an extra measurment (ULP plus)
-BSEC_SAMPLE_RATE_ULP_MEASUREMENT_ON_DEMAND = 0.0
-BSEC_SAMPLE_RATE_HIGH_PERFORMANCE = 0.055556
-
-# Measurement Value ranges
-# Min value for IAQ
-BSEC_MIN_IAQ = 0
-# Max value for IAQ
-BSEC_MAX_IAQ = 500
