--- backend/helpers/bme68x-python-library/examples/README.md
@@ -1,5 +1,0 @@
-Examples
-========
-
-All examples assume that you connected your BME68X sensor to I2C address 0x77 (119).<br>
-To use I2C address 0x76 (118) change the BME68X constructor call to BME68X_I2C_ADDR_LOW.
