--- backend/helpers/bme68x-python-library/tools/bmerawdata/default.bmeconfig
@@ -1,246 +1,0 @@
-{
-  "configHeader": {
-    "dateCreated": "2021-07-11T15:48:20.816Z",
-    "appVersion": "1.5.4",
-    "boardType": "board_8",
-    "boardMode": "heater_profile_exploration",
-    "boardLayout": "grouped"
-  },
-  "configBody": {
-    "heaterProfiles": [
-      {
-        "id": "heater_354",
-        "timeBase": 140,
-        "temperatureTimeVectors": [
-          [
-            320,
-            5
-          ],
-          [
-            100,
-            2
-          ],
-          [
-            100,
-            10
-          ],
-          [
-            100,
-            30
-          ],
-          [
-            200,
-            5
-          ],
-          [
-            200,
-            5
-          ],
-          [
-            200,
-            5
-          ],
-          [
-            320,
-            5
-          ],
-          [
-            320,
-            5
-          ],
-          [
-            320,
-            5
-          ]
-        ]
-      },
-      {
-        "id": "heater_301",
-        "timeBase": 140,
-        "temperatureTimeVectors": [
-          [
-            100,
-            2
-          ],
-          [
-            100,
-            41
-          ],
-          [
-            200,
-            2
-          ],
-          [
-            200,
-            14
-          ],
-          [
-            200,
-            14
-          ],
-          [
-            200,
-            14
-          ],
-          [
-            320,
-            2
-          ],
-          [
-            320,
-            14
-          ],
-          [
-            320,
-            14
-          ],
-          [
-            320,
-            14
-          ]
-        ]
-      },
-      {
-        "id": "heater_411",
-        "timeBase": 140,
-        "temperatureTimeVectors": [
-          [
-            100,
-            43
-          ],
-          [
-            320,
-            2
-          ],
-          [
-            170,
-            43
-          ],
-          [
-            320,
-            2
-          ],
-          [
-            240,
-            2
-          ],
-          [
-            240,
-            20
-          ],
-          [
-            240,
-            21
-          ],
-          [
-            320,
-            2
-          ],
-          [
-            320,
-            20
-          ],
-          [
-            320,
-            21
-          ]
-        ]
-      },
-      {
-        "id": "heater_501",
-        "timeBase": 140,
-        "temperatureTimeVectors": [
-          [
-            210,
-            24
-          ],
-          [
-            265,
-            2
-          ],
-          [
-            265,
-            22
-          ],
-          [
-            320,
-            2
-          ],
-          [
-            320,
-            22
-          ],
-          [
-            265,
-            24
-          ],
-          [
-            210,
-            24
-          ],
-          [
-            155,
-            24
-          ],
-          [
-            100,
-            24
-          ],
-          [
-            155,
-            24
-          ]
-        ]
-      }
-    ],
-    "dutyCycleProfiles": [
-      {
-        "id": "duty_3_6",
-        "numberScanningCycles": 3,
-        "numberSleepingCycles": 6
-      }
-    ],
-    "sensorConfigurations": [
-      {
-        "sensorIndex": 0,
-        "heaterProfile": "heater_354",
-        "dutyCycleProfile": "duty_3_6"
-      },
-      {
-        "sensorIndex": 1,
-        "heaterProfile": "heater_354",
-        "dutyCycleProfile": "duty_3_6"
-      },
-      {
-        "sensorIndex": 2,
-        "heaterProfile": "heater_301",
-        "dutyCycleProfile": "duty_3_6"
-      },
-      {
-        "sensorIndex": 3,
-        "heaterProfile": "heater_301",
-        "dutyCycleProfile": "duty_3_6"
-      },
-      {
-        "sensorIndex": 4,
-        "heaterProfile": "heater_411",
-        "dutyCycleProfile": "duty_3_6"
-      },
-      {
-        "sensorIndex": 5,
-        "heaterProfile": "heater_411",
-        "dutyCycleProfile": "duty_3_6"
-      },
-      {
-        "sensorIndex": 6,
-        "heaterProfile": "heater_501",
-        "dutyCycleProfile": "duty_3_6"
-      },
-      {
-        "sensorIndex": 7,
-        "heaterProfile": "heater_501",
-        "dutyCycleProfile": "duty_3_6"
-      }
-    ]
-  }
-}
