+++ logs/bme
@@ -1,0 +1,24 @@
+{
+    "sample_nr": 81,
+    "timestamp": 561005485970,
+    "iaq": 25.0,
+    "iaq_accuracy": 0,
+    "static_iaq": 25.0,
+    "static_iaq_accuracy": 0,
+    "co2_equivalent": 500.0,
+    "co2_accuracy": 0,
+    "breath_voc_equivalent": 0.4999999403953552,
+    "breath_voc_accuracy": 0,
+    "raw_temperature": 24.375450134277344,
+    "raw_pressure": 101824.7890625,
+    "raw_humidity": 38.6302490234375,
+    "raw_gas": 3971.53857421875,
+    "stabilization_status": 0,
+    "run_in_status": 0,
+    "temperature": 19.313644409179688,
+    "humidity": 52.635169982910156,
+    "comp_gas_value": 4.294103622436523,
+    "comp_gas_accuracy": 0,
+    "gas_percentage": 0.0,
+    "gas_percentage_accuracy": 0,
+}
